feat: FileInputField, move back to components
- updated template registry

- move onDeleteHandler type into delete-button

- updated tests

- updated docs
