+++ ember-toucan-core/src/components/form/file-input-field.hbs
