+++ ember-toucan-core/src/template-registry.ts
@@ -9,8 +9,8 @@ import type TextareaControlComponent from './components/form/controls/textarea';
 import type FieldComponent from './components/form/field';
 import type FieldsetComponent from './components/form/fieldset';
 import type FormFileInputDeleteButtonComponent from './components/form/file-input/delete-button';
-import type FormFileInputFieldComponent from './components/form/file-input/field';
 import type FormFileInputListComponent from './components/form/file-input/list';
+import type FormFileInputFieldComponent from './components/form/file-input-field';
 import type InputFieldComponent from './components/form/input-field';
 import type RadioFieldComponent from './components/form/radio-field';
 import type RadioGroupFieldComponent from './components/form/radio-group-field';
@@ -28,9 +28,9 @@ export default interface Registry {
   'Form::Controls::Input': typeof InputControlComponent;
   'Form::Controls::Radio': typeof RadioControlComponent;
   'Form::Controls::Textarea': typeof TextareaControlComponent;
-  'Form::FileInput::Field': typeof FormFileInputFieldComponent;
   'Form::FileInput::List': typeof FormFileInputListComponent;
   'Form::FileInput::DeleteButton': typeof FormFileInputDeleteButtonComponent;
+  'Form::FileInputField': typeof FormFileInputFieldComponent;
   'Form::RadioField': typeof RadioFieldComponent;
   'Form::RadioGroupField': typeof RadioGroupFieldComponent;
   'Form::TextareaField': typeof TextareaFieldComponent;
